Split files (which is just copy-paste)
